+++ src/SMAPI.Mods.ConsoleCommands/Framework/Commands/Player/ListItemsCommand.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using StardewModdingAPI.Mods.ConsoleCommands.Framework.ItemData;
 
 namespace StardewModdingAPI.Mods.ConsoleCommands.Framework.Commands.Player
 {
@@ -12,7 +11,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework.Commands.Player
         ** Fields
         *********/
         /// <summary>Provides methods for searching and constructing items.</summary>
-        private readonly ItemRepository Items = new ItemRepository();
+        private readonly ItemRepository Items = new();
 
 
         /*********
@@ -45,7 +44,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework.Commands.Player
                 .ToArray();
             string summary = "Searching...\n";
             if (matches.Any())
-                monitor.Log(summary + this.GetTableString(matches, new[] { "type", "name", "id" }, val => new[] { val.Type.ToString(), val.Name, val.ID.ToString() }), LogLevel.Info);
+                monitor.Log(summary + this.GetTableString(matches, new[] { "name", "id" }, val => new[] { val.Name, val.QualifiedItemID }), LogLevel.Info);
             else
                 monitor.Log(summary + "No items found", LogLevel.Info);
         }
@@ -66,7 +65,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework.Commands.Player
             // find matches
             return (
                 from item in this.Items.GetAll()
-                let term = $"{item.ID}|{item.Type}|{item.Name}|{item.DisplayName}"
+                let term = $"{item.QualifiedItemID}|{item.Type}|{item.Name}|{item.DisplayName}"
                 where searchWords == null || searchWords.All(word => term.IndexOf(word, StringComparison.CurrentCultureIgnoreCase) != -1)
                 select item
             );