+++ src/SMAPI.Mods.ConsoleCommands/Framework/ItemRepository.cs
@@ -4,7 +4,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
-using StardewModdingAPI.Mods.ConsoleCommands.Framework.ItemData;
 using StardewValley;
 using StardewValley.GameData.FishPond;
 using StardewValley.Menus;
@@ -18,20 +17,13 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
     internal class ItemRepository
     {
         /*********
-        ** Fields
-        *********/
-        /// <summary>The custom ID offset for items don't have a unique ID in the game.</summary>
-        private readonly int CustomIDOffset = 1000;
-
-
-        /*********
         ** Public methods
         *********/
         /// <summary>Get all spawnable items.</summary>
-        /// <param name="itemTypes">The item types to fetch (or null for any type).</param>
+        /// <param name="onlyType">Only include items for the given <see cref="ItemDataDefinition.Identifier"/>.</param>
         /// <param name="includeVariants">Whether to include flavored variants like "Sunflower Honey".</param>
         [SuppressMessage("ReSharper", "AccessToModifiedClosure", Justification = "TryCreate invokes the lambda immediately.")]
-        public IEnumerable<SearchableItem> GetAll(ItemType[] itemTypes = null, bool includeVariants = true)
+        public IEnumerable<SearchableItem> GetAll(string onlyType = null, bool includeVariants = true)
         {
             //
             //
@@ -45,145 +37,105 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
 
             IEnumerable<SearchableItem> GetAllRaw()
             {
-                HashSet<ItemType> types = itemTypes?.Any() == true ? new HashSet<ItemType>(itemTypes) : null;
-                bool ShouldGet(ItemType type) => types == null || types.Contains(type);
-
-                // get tools
-                if (ShouldGet(ItemType.Tool))
+                // get from item data definitions
+                foreach (ItemDataDefinition itemType in ItemDataDefinition.ItemTypes)
                 {
-                    for (int q = Tool.stone; q <= Tool.iridium; q++)
+                    if (onlyType != null && itemType.Identifier != onlyType)
+                        continue;
+
+                    switch (itemType.Identifier)
                     {
-                        int quality = q;
+                        // tools
+                        case "(T)":
+                            foreach (string id in itemType.GetAllItemIDs())
+                            {
+                                // base item
+                                SearchableItem baseTool = this.TryCreate(itemType.Identifier, id, p => itemType.CreateItem(p.ID, 1, SObject.lowQuality));
+                                yield return baseTool;
 
-                        yield return this.TryCreate(ItemType.Tool, ToolFactory.axe, _ => ToolFactory.getToolFromDescription(ToolFactory.axe, quality));
-                        yield return this.TryCreate(ItemType.Tool, ToolFactory.hoe, _ => ToolFactory.getToolFromDescription(ToolFactory.hoe, quality));
-                        yield return this.TryCreate(ItemType.Tool, ToolFactory.pickAxe, _ => ToolFactory.getToolFromDescription(ToolFactory.pickAxe, quality));
-                        yield return this.TryCreate(ItemType.Tool, ToolFactory.wateringCan, _ => ToolFactory.getToolFromDescription(ToolFactory.wateringCan, quality));
-                        if (quality != Tool.iridium)
-                            yield return this.TryCreate(ItemType.Tool, ToolFactory.fishingRod, _ => ToolFactory.getToolFromDescription(ToolFactory.fishingRod, quality));
+                                // handle tool upgrades
+                                if (baseTool.ID is nameof(Axe) or nameof(FishingRod) or nameof(Hoe) or nameof(Pickaxe) or nameof(WateringCan))
+                                {
+                                    for (int q = Tool.copper; q <= Tool.iridium; q++)
+                                    {
+                                        // avoid closure capture
+                                        int quality = q;
+
+                                        // no iridium fishing rod
+                                        if (quality == Tool.iridium && baseTool.ID == nameof(FishingRod))
+                                            continue;
+
+                                        // get upgraded tool
+                                        yield return this.TryCreate(itemType.Identifier, $"{baseTool.ID}/{quality}", _ =>
+                                        {
+                                            Tool tool = (Tool)baseTool.CreateItem();
+                                            tool.UpgradeLevel = quality;
+                                            return tool;
+                                        });
+                                    }
+                                }
+                            }
+                            break;
+
+                        // objects
+                        case "(O)":
+                            foreach (string id in itemType.GetAllItemIDs())
+                            {
+                                // base item
+                                SearchableItem result = this.TryCreate(itemType.Identifier, id, p => itemType.CreateItem(p.ID, 1, SObject.lowQuality));
+
+                                // ring
+                                if (result.Item is Ring)
+                                    yield return result;
+
+                                // journal scraps
+                                else if (result.QualifiedItemID == "(O)842")
+                                {
+                                    foreach (SearchableItem journalScrap in this.GetSecretNotes(itemType, isJournalScrap: true))
+                                        yield return journalScrap;
+                                }
+
+                                // secret notes
+                                else if (result.QualifiedItemID == "(O)79")
+                                {
+                                    foreach (SearchableItem secretNote in this.GetSecretNotes(itemType, isJournalScrap: false))
+                                        yield return secretNote;
+                                }
+
+                                // object
+                                else
+                                {
+                                    yield return result;
+
+                                    if (includeVariants)
+                                    {
+                                        foreach (SearchableItem variant in this.GetFlavoredObjectVariants((SObject)result.Item, itemType))
+                                            yield return variant;
+                                    }
+                                }
+                            }
+                            break;
+
+                        // no special handling needed
+                        default:
+                            foreach (string id in itemType.GetAllItemIDs())
+                                yield return this.TryCreate(itemType.Identifier, id, p => itemType.CreateItem(p.ID, 1, SObject.lowQuality));
+                            break;
                     }
-                    yield return this.TryCreate(ItemType.Tool, this.CustomIDOffset, _ => new MilkPail()); // these don't have any sort of ID, so we'll just assign some arbitrary ones
-                    yield return this.TryCreate(ItemType.Tool, this.CustomIDOffset + 1, _ => new Shears());
-                    yield return this.TryCreate(ItemType.Tool, this.CustomIDOffset + 2, _ => new Pan());
-                    yield return this.TryCreate(ItemType.Tool, this.CustomIDOffset + 3, _ => new Wand());
-                }
-
-                // clothing
-                if (ShouldGet(ItemType.Clothing))
-                {
-                    foreach (int id in this.GetShirtIds())
-                        yield return this.TryCreate(ItemType.Clothing, id, p => new Clothing(p.ID));
                 }
 
                 // wallpapers
-                if (ShouldGet(ItemType.Wallpaper))
+                if (onlyType is null or "(WP)")
                 {
                     for (int id = 0; id < 112; id++)
-                        yield return this.TryCreate(ItemType.Wallpaper, id, p => new Wallpaper(p.ID) { Category = SObject.furnitureCategory });
+                        yield return this.TryCreate("(WP)", id.ToString(), p => new Wallpaper(int.Parse(p.ID)) { Category = SObject.furnitureCategory });
                 }
 
                 // flooring
-                if (ShouldGet(ItemType.Flooring))
+                if (onlyType is null or "(FL)")
                 {
                     for (int id = 0; id < 56; id++)
-                        yield return this.TryCreate(ItemType.Flooring, id, p => new Wallpaper(p.ID, isFloor: true) { Category = SObject.furnitureCategory });
-                }
-
-                // equipment
-                if (ShouldGet(ItemType.Boots))
-                {
-                    foreach (int id in this.TryLoad<int, string>("Data\\Boots").Keys)
-                        yield return this.TryCreate(ItemType.Boots, id, p => new Boots(p.ID));
-                }
-                if (ShouldGet(ItemType.Hat))
-                {
-                    foreach (int id in this.TryLoad<int, string>("Data\\hats").Keys)
-                        yield return this.TryCreate(ItemType.Hat, id, p => new Hat(p.ID));
-                }
-
-                // weapons
-                if (ShouldGet(ItemType.Weapon))
-                {
-                    foreach (int id in this.TryLoad<int, string>("Data\\weapons").Keys)
-                    {
-                        yield return this.TryCreate(ItemType.Weapon, id, p => (p.ID >= 32 && p.ID <= 34)
-                            ? (Item)new Slingshot(p.ID)
-                            : new MeleeWeapon(p.ID)
-                        );
-                    }
-                }
-
-                // furniture
-                if (ShouldGet(ItemType.Furniture))
-                {
-                    foreach (int id in this.TryLoad<int, string>("Data\\Furniture").Keys)
-                        yield return this.TryCreate(ItemType.Furniture, id, p => Furniture.GetFurnitureInstance(p.ID));
-                }
-
-                // craftables
-                if (ShouldGet(ItemType.BigCraftable))
-                {
-                    foreach (int id in Game1.bigCraftablesInformation.Keys)
-                        yield return this.TryCreate(ItemType.BigCraftable, id, p => new SObject(Vector2.Zero, p.ID));
-                }
-
-                // objects
-                if (ShouldGet(ItemType.Object) || ShouldGet(ItemType.Ring))
-                {
-                    foreach (int id in Game1.objectInformation.Keys)
-                    {
-                        string[] fields = Game1.objectInformation[id]?.Split('/');
-
-                        // ring
-                        if (id != 801 && fields?.Length >= 4 && fields[3] == "Ring") // 801 = wedding ring, which isn't an equippable ring
-                        {
-                            if (ShouldGet(ItemType.Ring))
-                                yield return this.TryCreate(ItemType.Ring, id, p => new Ring(p.ID));
-                        }
-
-                        // journal scrap
-                        else if (id == 842)
-                        {
-                            if (ShouldGet(ItemType.Object))
-                            {
-                                foreach (SearchableItem journalScrap in this.GetSecretNotes(isJournalScrap: true))
-                                    yield return journalScrap;
-                            }
-                        }
-
-                        // secret notes
-                        else if (id == 79)
-                        {
-                            if (ShouldGet(ItemType.Object))
-                            {
-                                foreach (SearchableItem secretNote in this.GetSecretNotes(isJournalScrap: false))
-                                    yield return secretNote;
-                            }
-                        }
-
-                        // object
-                        else if (ShouldGet(ItemType.Object))
-                        {
-                            // spawn main item
-                            SObject item = null;
-                            yield return this.TryCreate(ItemType.Object, id, p =>
-                            {
-                                return item = (p.ID == 812 // roe
-                                    ? new ColoredObject(p.ID, 1, Color.White)
-                                    : new SObject(p.ID, 1)
-                                );
-                            });
-                            if (item == null)
-                                continue;
-
-                            // flavored items
-                            if (includeVariants)
-                            {
-                                foreach (SearchableItem variant in this.GetFlavoredObjectVariants(item))
-                                    yield return variant;
-                            }
-                        }
-                    }
+                        yield return this.TryCreate("(FL)", id.ToString(), p => new Wallpaper(int.Parse(p.ID), isFloor: true) { Category = SObject.furnitureCategory });
                 }
             }
 
@@ -195,12 +147,13 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
         ** Private methods
         *********/
         /// <summary>Get the individual secret note or journal scrap items.</summary>
+        /// <param name="itemType">The object data definition.</param>
         /// <param name="isJournalScrap">Whether to get journal scraps.</param>
         /// <remarks>Derived from <see cref="GameLocation.tryToCreateUnseenSecretNote"/>.</remarks>
-        private IEnumerable<SearchableItem> GetSecretNotes(bool isJournalScrap)
+        private IEnumerable<SearchableItem> GetSecretNotes(ItemDataDefinition itemType, bool isJournalScrap)
         {
             // get base item ID
-            int baseId = isJournalScrap ? 842 : 79;
+            string baseId = isJournalScrap ? "842" : "79";
 
             // get secret note IDs
             var ids = this
@@ -218,13 +171,9 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
             // build items
             foreach (int id in ids)
             {
-                int fakeId = this.CustomIDOffset * 8 + id;
-                if (isJournalScrap)
-                    fakeId += GameLocation.JOURNAL_INDEX;
-
-                yield return this.TryCreate(ItemType.Object, fakeId, _ =>
+                yield return this.TryCreate(itemType.Identifier, $"{baseId}/{id}", _ =>
                 {
-                    SObject note = new(baseId, 1);
+                    Item note = itemType.CreateItem(baseId, 1, SObject.lowQuality);
                     note.Name = $"{note.Name} #{id}";
                     return note;
                 });
@@ -233,16 +182,17 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
 
         /// <summary>Get flavored variants of a base item (like Blueberry Wine for Blueberry), if any.</summary>
         /// <param name="item">A sample of the base item.</param>
-        private IEnumerable<SearchableItem> GetFlavoredObjectVariants(SObject item)
+        /// <param name="itemType">The object data definition.</param>
+        private IEnumerable<SearchableItem> GetFlavoredObjectVariants(SObject item, ItemDataDefinition itemType)
         {
-            int id = item.ParentSheetIndex;
+            string id = item.ItemID;
 
             switch (item.Category)
             {
                 // fruit products
                 case SObject.FruitsCategory:
                     // wine
-                    yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 2 + id, _ => new SObject(348, 1)
+                    yield return this.TryCreate(itemType.Identifier, $"348/{id}", _ => new SObject("348", 1)
                     {
                         Name = $"{item.Name} Wine",
                         Price = item.Price * 3,
@@ -251,7 +201,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
                     });
 
                     // jelly
-                    yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 3 + id, _ => new SObject(344, 1)
+                    yield return this.TryCreate(itemType.Identifier, $"344/{id}", _ => new SObject("344", 1)
                     {
                         Name = $"{item.Name} Jelly",
                         Price = 50 + item.Price * 2,
@@ -263,7 +213,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
                 // vegetable products
                 case SObject.VegetableCategory:
                     // juice
-                    yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 4 + id, _ => new SObject(350, 1)
+                    yield return this.TryCreate(itemType.Identifier, $"350/{id}", _ => new SObject("350", 1)
                     {
                         Name = $"{item.Name} Juice",
                         Price = (int)(item.Price * 2.25d),
@@ -272,7 +222,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
                     });
 
                     // pickled
-                    yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 5 + id, _ => new SObject(342, 1)
+                    yield return this.TryCreate(itemType.Identifier, $"342/{id}", _ => new SObject("342", 1)
                     {
                         Name = $"Pickled {item.Name}",
                         Price = 50 + item.Price * 2,
@@ -283,9 +233,9 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
 
                 // flower honey
                 case SObject.flowersCategory:
-                    yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 5 + id, _ =>
+                    yield return this.TryCreate(itemType.Identifier, $"340/{id}", _ =>
                     {
-                        SObject honey = new SObject(Vector2.Zero, 340, $"{item.Name} Honey", false, true, false, false)
+                        SObject honey = new(Vector2.Zero, "340", $"{item.Name} Honey", false, true, false, false)
                         {
                             Name = $"{item.Name} Honey",
                             preservedParentSheetIndex = { id }
@@ -296,14 +246,14 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
                     break;
 
                 // roe and aged roe (derived from FishPond.GetFishProduce)
-                case SObject.sellAtFishShopCategory when id == 812:
+                case SObject.sellAtFishShopCategory when item.QualifiedItemID == "(O)812":
                     {
                         this.GetRoeContextTagLookups(out HashSet<string> simpleTags, out List<List<string>> complexTags);
 
                         foreach (var pair in Game1.objectInformation)
                         {
                             // get input
-                            SObject input = this.TryCreate(ItemType.Object, pair.Key, p => new SObject(p.ID, 1))?.Item as SObject;
+                            SObject input = this.TryCreate(itemType.Identifier, pair.Key, p => new SObject(p.ID, 1))?.Item as SObject;
                             var inputTags = input?.GetContextTags();
                             if (inputTags?.Any() != true)
                                 continue;
@@ -315,27 +265,27 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
                             // yield roe
                             SObject roe = null;
                             Color color = this.GetRoeColor(input);
-                            yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 7 + id, _ =>
+                            yield return this.TryCreate(itemType.Identifier, $"812/{input.ItemID}", _ =>
                             {
-                                roe = new ColoredObject(812, 1, color)
+                                roe = new ColoredObject("812", 1, color)
                                 {
                                     name = $"{input.Name} Roe",
                                     preserve = { Value = SObject.PreserveType.Roe },
-                                    preservedParentSheetIndex = { Value = input.ParentSheetIndex }
+                                    preservedParentSheetIndex = { Value = input.ItemID }
                                 };
                                 roe.Price += input.Price / 2;
                                 return roe;
                             });
 
                             // aged roe
-                            if (roe != null && pair.Key != 698) // aged sturgeon roe is caviar, which is a separate item
+                            if (roe != null && pair.Key != "698") // aged sturgeon roe is caviar, which is a separate item
                             {
-                                yield return this.TryCreate(ItemType.Object, this.CustomIDOffset * 7 + id, _ => new ColoredObject(447, 1, color)
+                                yield return this.TryCreate(itemType.Identifier, $"447/{input.ItemID}", _ => new ColoredObject("447", 1, color)
                                 {
                                     name = $"Aged {input.Name} Roe",
                                     Category = -27,
                                     preserve = { Value = SObject.PreserveType.AgedRoe },
-                                    preservedParentSheetIndex = { Value = input.ParentSheetIndex },
+                                    preservedParentSheetIndex = { Value = input.ItemID },
                                     Price = roe.Price * 2
                                 });
                             }
@@ -355,7 +305,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
 
             foreach (FishPondData data in Game1.content.Load<List<FishPondData>>("Data\\FishPondData"))
             {
-                if (data.ProducedItems.All(p => p.ItemID != 812))
+                if (data.ProducedItems.All(p => p.ItemID != "812"))
                     continue; // doesn't produce roe
 
                 if (data.RequiredTags.Count == 1 && !data.RequiredTags[0].StartsWith("!"))
@@ -384,13 +334,13 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
 
         /// <summary>Create a searchable item if valid.</summary>
         /// <param name="type">The item type.</param>
-        /// <param name="id">The unique ID (if different from the item's parent sheet index).</param>
+        /// <param name="key">The locally unique item key.</param>
         /// <param name="createItem">Create an item instance.</param>
-        private SearchableItem TryCreate(ItemType type, int id, Func<SearchableItem, Item> createItem)
+        private SearchableItem TryCreate(string type, string key, Func<SearchableItem, Item> createItem)
         {
             try
             {
-                var item = new SearchableItem(type, id, createItem);
+                var item = new SearchableItem(type, key, createItem);
                 item.Item.getDescription(); // force-load item data, so it crashes here if it's invalid
                 return item;
             }
@@ -405,47 +355,9 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework
         /// <remarks>Derived from <see cref="StardewValley.Buildings.FishPond.GetFishProduce"/>.</remarks>
         private Color GetRoeColor(SObject fish)
         {
-            return fish.ParentSheetIndex == 698 // sturgeon
+            return fish.QualifiedItemID == "(O)698" // sturgeon
                 ? new Color(61, 55, 42)
                 : (TailoringMenu.GetDyeColor(fish) ?? Color.Orange);
-        }
-
-        /// <summary>Get valid shirt IDs.</summary>
-        /// <remarks>
-        /// Shirts have a possible range of 1000–1999, but not all of those IDs are valid. There are two sets of IDs:
-        ///
-        /// <list type="number">
-        ///   <item>
-        ///     Shirts which exist in <see cref="Game1.clothingInformation"/>.
-        ///   </item>
-        ///   <item>
-        ///     Shirts with a dynamic ID and no entry in <see cref="Game1.clothingInformation"/>. These automatically
-        ///     use the generic shirt entry with ID <c>-1</c> and are mapped to a calculated position in the
-        ///     <c>Characters/Farmer/shirts</c> spritesheet. There's no constant we can use, but some known valid
-        ///     ranges are 1000–1111 (used in <see cref="Farmer.changeShirt"/> for the customization screen and
-        ///     1000–1127 (used in <see cref="Utility.getShopStock"/> and <see cref="GameLocation.sandyShopStock"/>).
-        ///     Based on the spritesheet, the max valid ID is 1299.
-        ///   </item>
-        /// </list>
-        /// </remarks>
-        private IEnumerable<int> GetShirtIds()
-        {
-            // defined shirt items
-            foreach (int id in Game1.clothingInformation.Keys)
-            {
-                if (id < 0)
-                    continue; // placeholder data for character customization clothing below
-
-                yield return id;
-            }
-
-            // dynamic shirts
-            HashSet<int> clothingIds = new HashSet<int>(Game1.clothingInformation.Keys);
-            for (int id = 1000; id <= 1299; id++)
-            {
-                if (!clothingIds.Contains(id))
-                    yield return id;
-            }
         }
     }
 }